# docs/reference/PharmaScope_Lite_v1.0/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PharmaScope Lite API
  version: 1.0.0
  description: Implemented PharmaScope Lite API. All IDs are workspace scoped and authenticated. Replay
    mode uses fictional fixtures; live mode requires PostgreSQL, source access and model credentials.
    Evidence references are validated server-side; model execution status is not implied by this contract.
servers:
- url: /
  description: Same-origin API behind the deployment reverse proxy
security:
- sessionCookie: []
paths:
...
  /api/v1/workspaces/{workspace_id}/drugs:
    get:
      operationId: drugs_list
      summary: drugs_list
      tags:
      - drugs
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - *id001
      - *id002
      - name: query
        in: query
        required: false
        schema:
          type: string
      - name: archived
        in: query
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrugPage'
        '400':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - sessionCookie: []
    post:
      operationId: drugs_create
      summary: drugs_create
      tags:
      - drugs
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
          minLength: 1
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drug'
        '400':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DrugInput'
...
  /api/v1/workspaces/{workspace_id}/trials:
    get:
      operationId: trials_list
      summary: trials_list
      tags:
      - trials
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - *id001
      - *id002
      - name: drug_id
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Id'
      - name: query
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: has_results
        in: query
        required: false
        schema:
          type: boolean
      - name: observed_since
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrialPage'
        '400':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - sessionCookie: []
...
  /api/v1/workspaces/{workspace_id}/publications:
    get:
      operationId: publications_list
      summary: publications_list
      tags:
      - publications
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - *id001
      - *id002
      - name: drug_id
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Id'
      - name: query
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: has_results
        in: query
        required: false
        schema:
          type: boolean
      - name: observed_since
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicationPage'
        '400':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - sessionCookie: []
...
  /api/v1/workspaces/{workspace_id}/events:
    get:
      operationId: events_list
      summary: events_list
      tags:
      - events
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - *id001
      - *id002
      - name: drug_id
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Id'
      - name: observed_since
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPage'
        '400':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - sessionCookie: []
...
  /api/v1/workspaces/{workspace_id}/reports:
    get:
      operationId: reports_list
      summary: reports_list
      tags:
      - reports
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - *id001
      - *id002
      - name: state
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportPage'
        '400':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - sessionCookie: []
...
  /api/v1/workspaces/{workspace_id}/inbox:
    get:
      operationId: inbox_list
      summary: inbox_list
      tags:
      - inbox
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - *id001
      - *id002
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeliveryPage'
        '400':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Structured error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - sessionCookie: []
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pharma_session
  schemas:
    Id:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              minLength: 1
            message:
              type: string
              minLength: 1
            details:
              type: object
              properties: {}
              required: []
              additionalProperties: true
            request_id:
              type: string
              minLength: 1
          required:
          - code
          - message
          - details
          - request_id
          additionalProperties: false
      required:
      - error
      additionalProperties: false
...
    DrugInput:
      type: object
      properties: &id043
        display_name: &id004
          type: string
          minLength: 1
          maxLength: 200
        development_code: &id005
          anyOf:
          - type: string
            minLength: 1
            maxLength: 100
          - type: 'null'
        description: &id006
          anyOf:
          - type: string
            maxLength: 3000
          - type: 'null'
        indications: &id007
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 200
          maxItems: 20
        targets: &id008
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 200
          maxItems: 20
      required:
      - display_name
      additionalProperties: false
    Drug:
      type: object
      properties:
        id: &id009
          $ref: '#/components/schemas/Id'
        workspace_id: &id010
          $ref: '#/components/schemas/Id'
        created_at: &id011
          $ref: '#/components/schemas/Timestamp'
        display_name: *id004
        development_code: *id005
        description: *id006
        indications: *id007
        targets: *id008
        revision:
          type: integer
          minimum: 1
        archived:
          type: boolean
        updated_at:
          $ref: '#/components/schemas/Timestamp'
      required:
      - id
      - workspace_id
      - created_at
      - display_name
      - development_code
      - description
      - indications
      - targets
      - revision
      - archived
      - updated_at
      additionalProperties: false
...
    DrugPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Drug'
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
        has_more:
          type: boolean
      required:
      - items
      - next_cursor
      - has_more
      additionalProperties: false
...
    TrialPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Trial'
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
        has_more:
          type: boolean
      required:
      - items
      - next_cursor
      - has_more
      additionalProperties: false
    PublicationPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Publication'
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
        has_more:
          type: boolean
      required:
      - items
      - next_cursor
      - has_more
      additionalProperties: false
...
    EventPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
        has_more:
          type: boolean
      required:
      - items
      - next_cursor
      - has_more
      additionalProperties: false
...